separate map components
